chore: rename reducer
